requirements.txt
google-genai
python-dotenv
pypdf
faiss-cpu
sentence-transformers
langchain-text-splitters